Rename "async_jobs.py" to "async_requests.py"
These asynchronous requests are no longer "jobs" in the girder sense.
They are more like asynchronous http requests. So name it such.
